+++ pkg/volume/emptydir/empty_dir.go
@@ -30,7 +30,7 @@ import (
 	"k8s.io/kubernetes/pkg/util/mount"
 	"k8s.io/kubernetes/pkg/volume"
 	volumeutil "k8s.io/kubernetes/pkg/volume/util"
-	"k8s.io/kubernetes/pkg/volume/util/quota"
+	"k8s.io/kubernetes/pkg/volume/util/fsquota"
 	utilstrings "k8s.io/utils/strings"
 )
 
@@ -241,12 +241,12 @@ func (ed *emptyDir) SetUpAt(dir string, mounterArgs volume.MounterArgs) error {
 		if mounterArgs.DesiredSize != nil {
 			// Deliberately shadow the outer use of err as noted
 			// above.
-			hasQuotas, err := quota.SupportsQuotas(ed.mounter, dir)
+			hasQuotas, err := fsquota.SupportsQuotas(ed.mounter, dir)
 			if err != nil {
 				klog.V(3).Infof("Unable to check for quota support on %s: %s", dir, err.Error())
 			} else if hasQuotas {
 				klog.V(4).Infof("emptydir trying to assign quota %v on %s", mounterArgs.DesiredSize, dir)
-				err := quota.AssignQuota(ed.mounter, dir, mounterArgs.PodUID, mounterArgs.DesiredSize)
+				err := fsquota.AssignQuota(ed.mounter, dir, mounterArgs.PodUID, mounterArgs.DesiredSize)
 				if err != nil {
 					klog.V(3).Infof("Set quota on %s failed %s", dir, err.Error())
 				}
@@ -416,7 +416,7 @@ func (ed *emptyDir) TearDownAt(dir string) error {
 
 func (ed *emptyDir) teardownDefault(dir string) error {
 	// Remove any quota
-	err := quota.ClearQuota(ed.mounter, dir)
+	err := fsquota.ClearQuota(ed.mounter, dir)
 	if err != nil {
 		klog.Warningf("Warning: Failed to clear quota on %s: %v", dir, err)
 	}
