+++ pkg/volume/util/fsquota/quota.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package quota
+package fsquota
 
 import (
 	"k8s.io/apimachinery/pkg/api/resource"
